requirements.txt
anyio==3.7.0
async-timeout==4.0.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-telegram-bot==20.3
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
